rename router to routes
